Update package directory organization
