--- pakmanaged.js
@@ -1,118 +1,0 @@
-var global = Function("return this;")();
-/*!
-  * Ender: open module JavaScript framework (client-lib)
-  * copyright Dustin Diaz & Jacob Thornton 2011 (@ded @fat)
-  * http://ender.no.de
-  * License MIT
-  */
-!function (context) {
-
-  // a global object for node.js module compatiblity
-  // ============================================
-
-  context['global'] = context
-
-  // Implements simple module system
-  // losely based on CommonJS Modules spec v1.1.1
-  // ============================================
-
-  var modules = {}
-    , old = context.$
-
-  function require (identifier) {
-    // modules can be required from ender's build system, or found on the window
-    var module = modules[identifier] || window[identifier]
-    if (!module) throw new Error("Requested module '" + identifier + "' has not been defined.")
-    return module
-  }
-
-  function provide (name, what) {
-    return (modules[name] = what)
-  }
-
-  context['provide'] = provide
-  context['require'] = require
-
-  function aug(o, o2) {
-    for (var k in o2) k != 'noConflict' && k != '_VERSION' && (o[k] = o2[k])
-    return o
-  }
-
-  function boosh(s, r, els) {
-    // string || node || nodelist || window
-    if (typeof s == 'string' || s.nodeName || (s.length && 'item' in s) || s == window) {
-      els = ender._select(s, r)
-      els.selector = s
-    } else els = isFinite(s.length) ? s : [s]
-    return aug(els, boosh)
-  }
-
-  function ender(s, r) {
-    return boosh(s, r)
-  }
-
-  aug(ender, {
-      _VERSION: '0.3.6'
-    , fn: boosh // for easy compat to jQuery plugins
-    , ender: function (o, chain) {
-        aug(chain ? boosh : ender, o)
-      }
-    , _select: function (s, r) {
-        return (r || document).querySelectorAll(s)
-      }
-  })
-
-  aug(boosh, {
-    forEach: function (fn, scope, i) {
-      // opt out of native forEach so we can intentionally call our own scope
-      // defaulting to the current item and be able to return self
-      for (i = 0, l = this.length; i < l; ++i) i in this && fn.call(scope || this[i], this[i], i, this)
-      // return self for chaining
-      return this
-    },
-    $: ender // handy reference to self
-  })
-
-  ender.noConflict = function () {
-    context.$ = old
-    return this
-  }
-
-  if (typeof module !== 'undefined' && module.exports) module.exports = ender
-  // use subscript notation as extern for Closure compilation
-  context['ender'] = context['$'] = context['ender'] || ender
-
-}(this);
-// pakmanager:wilson-interval
-(function (context) {
-  
-  var module = { exports: {} }, exports = module.exports
-    , $ = require("ender")
-    ;
-  
-  module.exports = function(ups, downs, z) {
-    		
-    	var n = ups + downs;
-    	
-    	if (n <= 0 || n < ups) return 0;
-    			
-    	var z = z || 1.036 // 1.036 = 85%, 1.645 = 95%
-    	
-    	var p = (ups/n);
-    	
-    	var left = (1/(1+((1/n)*(z*z))));
-    	
-    	var middle = (p+((1/(2*n))*(z*z)))
-    	
-    	var diff = (z * Math.sqrt(((1/n)*p*(1-p))+((1/(4*(n*n)))*(z*z))));
-    	
-    	var wilson = {
-    		high: (left * (middle + diff)),
-    		center: ((p + ((1/(2*n))*(z*z))) / (1 + ((1/n)*(z*z)))),
-    		low: (left * (middle - diff))
-    	};
-    	
-    	return wilson;
-    }
-  provide("wilson-interval", module.exports);
-}(global));
